Update connect.py
There is a disconnect between the intention of updating Thomas' age and the execution.

Thomas starts with the age of 44 and the intention of this demo is to change the age to something else.  Unfortunately, the code as it is, changes the age to 44 (which is no change at all.)

Also.... the commentary says that TWO ages are being changed but actually only one is.  So I changed the code to reflect that only one has been changed.

I propose to change it to "100" which has different digits from 44 and three digits so it makes it extra clear that 44 is being changed to something different.